backend/requirements.txt
aistudio-sdk==0.3.8
albucore==0.0.13
albumentations==1.4.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astor==0.8.1
bce-python-sdk==0.9.68
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
chardet==7.4.0.post2
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.2
crc32c==2.8
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
einops==0.2.0
exceptiongroup==1.3.1
fastapi==0.135.2
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
fsspec==2026.3.0
future==1.0.0
gdown==4.4.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
ImageIO==2.37.3
imagesize==2.0.0
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
lmdb==2.2.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.35.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.9.1
paddlepaddle==2.6.2
paddlex==3.4.3
pandas==2.3.3
pillow==10.2.0
pip==23.0.1
prefetch-generator==1.0.1
prettytable==3.17.0
protobuf==3.20.2
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pypdfium2==5.6.0
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.2
RapidFuzz==3.14.3
requests==2.33.0
rich==14.3.3
ruamel.yaml==0.19.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
setuptools==65.5.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
sympy==1.14.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.4.1
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.42.0
vietocr==0.3.13
wcwidth==0.6.0
Werkzeug==3.1.7